refactor: change ias structure
Remove all infrastructure dependencies and runners into the root path.
That change was done to prepare the application testing structure for
E2E, integrated tests, and stack stress testing.
