+++ .eslintrc.js
@@ -15,13 +15,12 @@ module.exports = {
     },
   ],
   ignorePatterns: [
-    'dist/**',
-    'build/**',
-    'bin/**',
+    '**/dist/**',
+    '**/build/**',
+    '**/bin/**',
     'templates/**',
     '*.py',
-    'cdk.out/**',
-    'api/dist/**',
+    '**/cdk.out/**',
     '.eslintrc.js',
   ],
   parser: '@typescript-eslint/parser',
