+++ .husky/post-merge
@@ -1,4 +1,8 @@
 #!/bin/sh
 . "$(dirname "$0")/_/husky.sh"
 
-npm ci
+npm ci \\
+&& cd ./infrastructure \\
+&& npm ci \\
+&& cd ../api \\
+&& npm ci && cd ..
