+++ dependabot.yml
@@ -11,6 +11,17 @@ updates:
       - 'npm'
       - 'dependencies'
       - 'security'
+  - package-ecosystem: 'npm'
+    directory: '/infrastructure'
+    schedule:
+      interval: 'daily'
+    commit-message:
+      prefix: 'chore:'
+      include: 'scope'
+    labels:
+      - 'npm'
+      - 'dependencies'
+      - 'security'
       - 'infrastructure'
   - package-ecosystem: 'npm'
     directory: '/api'
