+++ infrastructure/cdk.json
@@ -1,5 +1,5 @@
 {
-  "app": "npx ts-node --prefer-ts-exts infrastructure/bin.ts",
+  "app": "npx ts-node --prefer-ts-exts src/bin.ts",
   "watch": {
     "include": ["**"],
     "exclude": [
