+++ infrastructure/jest.config.js
@@ -1,6 +1,6 @@
 module.exports = {
   testEnvironment: 'node',
-  roots: ['<rootDir>/test'],
+  roots: ['<rootDir>/src'],
   testMatch: ['**/*.spec.ts'],
   transform: {
     '^.+\\.tsx?$': 'ts-jest',
