+++ infrastructure/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "todo-ias-ts-infra",
+  "version": "0.1.0",
+  "bin": {
+    "ias": "./dist/infrastructure/bin.js"
+  },
+  "scripts": {
+    "start:local": "sam local start-api  -t ./cdk.out/CdkTaStack.template.json",
+    "build": "tsc",
+    "watch": "tsc -w",
+    "test": "jest",
+    "cdk": "cdk",
+    "test:debug": "node --inspect-brk ./node_modules/jest/bin/jest.js --runInBand --coverage=false"
+  },
+  "devDependencies": {
+    "@commitlint/cli": "17.0.3",
+    "@commitlint/config-angular": "17.0.3",
+    "@commitlint/config-conventional": "17.0.3",
+    "@types/jest": "28.1.4",
+    "@types/node": "16.11.43",
+    "aws-cdk": "2.30.0",
+    "jest": "28.1.2",
+    "ts-jest": "28.0.5",
+    "ts-node": "10.8.2",
+    "typescript": "4.7.4"
+  },
+  "dependencies": {
+    "aws-cdk-lib": "2.30.0",
+    "constructs": "10.1.43",
+    "nanoid": "4.0.0",
+    "source-map-support": "0.5.21"
+  }
+}
