+++ infrastructure/src/bin.ts
