+++ infrastructure/src/resource/dynamo-db.ts
