+++ infrastructure/src/stack-resource.ts
