+++ infrastructure/tsconfig.json
@@ -25,5 +25,5 @@
     "esModuleInterop": true,
     "typeRoots": ["./node_modules/@types"]
   },
-  "exclude": ["node_modules", "infrastructure/cdk.out", "api", "**/*.spec.ts"]
+  "exclude": ["node_modules", "./cdk.out", "**/*.spec.ts"]
 }
