+++ package.json
@@ -5,12 +5,6 @@
     "ias": "./dist/infrastructure/bin.js"
   },
   "scripts": {
-    "start:local": "sam local start-api  -t ./cdk.out/CdkTaStack.template.json",
-    "build": "tsc",
-    "watch": "tsc -w",
-    "test": "jest",
-    "cdk": "cdk",
-    "test:debug": "node --inspect-brk ./node_modules/jest/bin/jest.js --runInBand --coverage=false",
     "lint:commit:format": "npx commitlint --from HEAD~1 --to HEAD --verbose",
     "lint:code:format": "prettier .",
     "lint:code:format:fix": "prettier --write .",
@@ -21,26 +15,18 @@
     "@commitlint/cli": "17.0.3",
     "@commitlint/config-angular": "17.0.3",
     "@commitlint/config-conventional": "17.0.3",
-    "@types/jest": "28.1.4",
     "@types/node": "16.11.43",
-    "aws-cdk": "2.30.0",
     "husky": "8.0.1",
-    "jest": "28.1.2",
-    "ts-jest": "28.0.5",
     "ts-node": "10.8.2",
     "typescript": "4.7.4"
   },
   "dependencies": {
     "@typescript-eslint/eslint-plugin": "5.30.5",
     "@typescript-eslint/parser": "5.30.5",
-    "aws-cdk-lib": "2.30.0",
-    "constructs": "10.1.43",
     "eslint": "8.19.0",
     "eslint-config-prettier": "8.5.0",
     "eslint-plugin-prettier": "4.2.1",
     "eslint-plugin-unused-imports": "2.0.0",
-    "moment": "2.29.4",
-    "nanoid": "4.0.0",
     "prettier": "2.7.1",
     "source-map-support": "0.5.21"
   }
